Split config manager to a separate file
